ISSUE: Corrected a "Segmentation fault" reported by Shengfeng Huang <hshengf2@mail.sysu.edu.cn> on 2012/9/24. He sent an email with a xxxx.bam file that produced the error.
PROBLEM:
Original bam2hints created the following arrays with a fixed size of "10"

  int PSLb[10]; int PSLq[10]; int PSLt[10]; 
  int BlockBegins[10]; int BlockEnds[10]; bool FolIntOK[10];   

Whereby these arrays hold information extrated from the CIGAR string.

The file xxxx.bam contains alignments with relatively long CIGARs and the capacity of the 
aforementioned arrays was exceeded.

The size of these arrays does not exceed the value iside the variable "block".

SOLUTION:
Created a new routine that estimates the maximum observered "block" within the complete BAM file.
The value of this new variable "maxBlock" is used to create the arrays.

Tonatiuh




git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@460 946bced8-b911-4d30-a262-e0a769a8c5ad

Former-commit-id: 19db9064532c791c20995921c099c2ea08de68f9