bam2wig/Makefile: Fix `TOOLDIR` handling to match HTSlib build prefix
HTSlib uses GNU Autotools, and the prefix defaults to `/usr/local`.
There, the headers are installed into the directory `include/htslib` and
the libraries into `lib`.

The current handling for some reason seems to install everything into a
directory `htslib`.

Change the handling, so that `TOOLDIR` matches the used prefix, when
building HTSlib.